Improved Python scripts and cv32e40p Makefile.

+++ checks/get_jg_summary.py
@@ -1,3 +1,6 @@
+import os
+
+# ====================  FUNCTIONS BEGIN  ==================== #
 
 def get_summary_from_log():
     vec = []
@@ -17,18 +20,24 @@ def get_summary_from_log():
                     no_asserts = line.split()[-1]
                 if "- proven" in line:
                     no_proven = line.split()[-1]
-    if no_asserts != no_proven:
+    if no_asserts != no_proven or found_summary == False:
         all_proven = False
     return vec, all_proven
 
+# ====================  FUNCTIONS END  ==================== #
 
+# ====================  MAIN CODE BEGIN  ==================== #  
 
-summary_vec, success = get_summary_from_log()
-if success:
-    log_name = "SUCCESS.log"
+if os.path.exists('jgproject/jg.log'):
+    summary_vec, success = get_summary_from_log()
+    if success:
+        log_name = "SUCCESS.log"
+    else:
+        log_name = "FAIL.log"
+    with open(log_name, 'w') as file:
+        for line in summary_vec:
+            file.write(line)
 else:
-    log_name = "FAIL.log"
-with open(log_name, 'w') as file:
-    for line in summary_vec:
-        file.write(line)
-    
+    print("Cannot find jgproject/jg.log")
+
+# ====================  MAIN CODE END  ==================== #
